backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
Flask>=3.0.0
flask-cors>=5.0.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
scikit-learn>=1.5.0
spacy>=3.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
PyMuPDF>=1.24.0
numpy>=1.26.0
requests>=2.31.0
gunicorn>=21.2.0
python-docx>=1.1.0
